% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_vaso_data.R
\name{delete_vaso_data}
\alias{delete_vaso_data}
\title{Delete VASO Image Filenames}
\usage{
delete_vaso_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{delete_kirby21_data}},
\code{modalities = "VASO"} so it cannot be specified}
}
\value{
Nothing is returned
}
\description{
Delete the files for the VASO images
}
