% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vaso_data.R
\name{download_vaso_data}
\alias{download_vaso_data}
\title{Download VASO Image Filenames}
\usage{
download_vaso_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{download_kirby21_data}},
\code{modalities = "VASO"} so it cannot be specified}
}
\value{
Indicator if the data is downloaded.
}
\description{
Download the files for the VASO images
}
\examples{
download_vaso_data()
}
